Game State System Rework & Misc. Changes
Reworked the game state system so that it directly uses the Class of a
game state instead of an enum, this makes it take less code to add a new
game state.  Also added a simple-ish way to pass arguments to a game
state's setup function (not tested yet).  Renamed Renderer to prepare
for a future graphics layering system rework.  Some other miscellaneous
fixes, cleanup, etc.