Graphics Layering System Rework and Misc. Changes
Reworked the graphics layering system so that you can create an instance
of a "LayerSet". Renderer is now a simpler interface for PrimaryRenderer
(a slightly cleaner way for working with Renderers drawing directly to a
main layer), LayerSet, and any other subclasses.  Separated out what was
originally the main LayerContainer into MainLayerSet (thin layer over
LayerSet) and MainLayerSetContainer (the AWT/Swing stuff and contains a
MainLayerSet).  Some other miscelaneous fixes and changes, like fixing a
BGM bug when resuming, and the player unit is now a demo of the
animation-path-changing setup.  Also a few new sample classes for new
stuff.