Implemented Game Clocks
Implemented a Clock class which is used by game states and anything else
to pause, resume, and do any kind of timing-based things.  Added a Clock
instance that is connected to the game states (reset when the game state
is changed, etc.).  Also renamed the ProgramTimer to ProgramClock and
made some small setup changes to game states.