Various Cleanup and Removal of "Singletons"
Moved around some of the setup operations. Renamed SetupOperations to
EngineSetupData and moved it to a different package. Changed input
devices from "Singleton" (single instance) setup to a static access
setup. Replaced many uses of synchronized in function declarations to
block synchronizing on objects. Added a method to layer sets to add a
renderer to all layers in the set. Also other miscellaneous cleanup.